[CHORE]#245 : Move Sender Pacakage Name
