+++ data-remote/src/main/kotlin/com/bff/wespot/data/remote/model/message/response/SenderDto.kt
@@ -1,6 +1,6 @@
 package com.bff.wespot.data.remote.model.message.response
 
-import com.bff.wespot.model.message.Sender
+import com.bff.wespot.model.message.response.Sender
 import kotlinx.serialization.Serializable
 
 @Serializable
